Final overhaul of entire project as previous implementation was not able to be adapted well for Assignment 2:
Populated Creature with abstract methods
Added Overrides to Player.cs
implemented the logic for movement between three rooms
Made use of inheritance throughout
Made use of a LINQ statement + Lambda expression to alphabetically sort inventory items
Implemented enemies which are set within rooms to attack the player, when the player attacks them. This can kill the player if they are not careful.
implemented a wider inventory that the player can look through and use, with varying functionality.
Tried to make the code encapsulated and modular throughout, worked on ensuring the classes used were abstracted as best I could do with time constraints